Feature: UserCard component finished.
But unsolved error in sub component. It will be fixed in the future.
